+++ src/_demo_syslog.jl
@@ -6,14 +6,15 @@
 # Edit src/sysstate.yaml instead
 
 """
-    demo_syslog(P, O=1, D=0, L=0, W=1, T=W, S=0; duration=10)
+    demo_syslog(P, O=1, D=0, L=0, W=1, T=W, S=0, N=0; duration=10)
 
 Create a demo flight log  with given duration [s] as StructArray. P is the number of tether
 particles, O the number of oriented frames (demo data fills only frame 1), D the
 number of aero segments (demo data has no flaps, so D=0), L the number of
-pulleys (demo data has none, so L=0) and S the number of segments (likewise 0).
+pulleys (demo data has none, so L=0), S the number of segments (likewise 0) and
+N the number of aerodynamic panels (likewise 0).
 """
-function demo_syslog(P, O=1, D=0, L=0, W=1, T=W, S=0; duration=10)
+function demo_syslog(P, O=1, D=0, L=0, W=1, T=W, S=0, N=0; duration=10)
     max_height = 6.03
     steps   = Int(duration * se().sample_freq) + 1
     time_vec = Vector{Float64}(undef, steps)
@@ -52,10 +53,18 @@ function demo_syslog(P, O=1, D=0, L=0, W=1, T=W, S=0; duration=10)
     alpha4_vec = Vector{MyFloat}(undef, steps)
     CL2_vec = Vector{MyFloat}(undef, steps)
     CD2_vec = Vector{MyFloat}(undef, steps)
-    aero_force_b_vec = Vector{MVector{3, MyFloat}}(undef, steps)
-    aero_moment_b_vec = Vector{MVector{3, MyFloat}}(undef, steps)
-    tether_induced_force_vec = Vector{MVector{3, MyFloat}}(undef, steps)
-    tether_induced_moment_vec = Vector{MVector{3, MyFloat}}(undef, steps)
+    aero_force_b_x_vec = Vector{MVector{O, MyFloat}}(undef, steps)
+    aero_force_b_y_vec = Vector{MVector{O, MyFloat}}(undef, steps)
+    aero_force_b_z_vec = Vector{MVector{O, MyFloat}}(undef, steps)
+    aero_moment_b_x_vec = Vector{MVector{O, MyFloat}}(undef, steps)
+    aero_moment_b_y_vec = Vector{MVector{O, MyFloat}}(undef, steps)
+    aero_moment_b_z_vec = Vector{MVector{O, MyFloat}}(undef, steps)
+    tether_induced_force_b_x_vec = Vector{MVector{O, MyFloat}}(undef, steps)
+    tether_induced_force_b_y_vec = Vector{MVector{O, MyFloat}}(undef, steps)
+    tether_induced_force_b_z_vec = Vector{MVector{O, MyFloat}}(undef, steps)
+    tether_induced_moment_b_x_vec = Vector{MVector{O, MyFloat}}(undef, steps)
+    tether_induced_moment_b_y_vec = Vector{MVector{O, MyFloat}}(undef, steps)
+    tether_induced_moment_b_z_vec = Vector{MVector{O, MyFloat}}(undef, steps)
     twist_angles_vec = Vector{MVector{D, MyFloat}}(undef, steps)
     vel_kite_vec = Vector{MVector{3, MyFloat}}(undef, steps)
     acc_vec = Vector{MyFloat}(undef, steps)
@@ -73,6 +82,7 @@ function demo_syslog(P, O=1, D=0, L=0, W=1, T=W, S=0; duration=10)
     drag_force_y_vec = Vector{MVector{P, MyFloat}}(undef, steps)
     drag_force_z_vec = Vector{MVector{P, MyFloat}}(undef, steps)
     spring_force_vec = Vector{MVector{S, MyFloat}}(undef, steps)
+    gamma_distribution_vec = Vector{MVector{N, MyFloat}}(undef, steps)
     turn_rate_x_vec = Vector{MVector{O, MyFloat}}(undef, steps)
     turn_rate_y_vec = Vector{MVector{O, MyFloat}}(undef, steps)
     turn_rate_z_vec = Vector{MVector{O, MyFloat}}(undef, steps)
@@ -82,6 +92,9 @@ function demo_syslog(P, O=1, D=0, L=0, W=1, T=W, S=0; duration=10)
     set_torque_vec = Vector{MVector{W, MyFloat}}(undef, steps)
     set_speed_vec = Vector{MVector{W, MyFloat}}(undef, steps)
     set_force_vec = Vector{MVector{W, MyFloat}}(undef, steps)
+    set_ext_force_enu_x_vec = Vector{MVector{P, MyFloat}}(undef, steps)
+    set_ext_force_enu_y_vec = Vector{MVector{P, MyFloat}}(undef, steps)
+    set_ext_force_enu_z_vec = Vector{MVector{P, MyFloat}}(undef, steps)
     var_01_vec = Vector{MyFloat}(undef, steps)
     var_02_vec = Vector{MyFloat}(undef, steps)
     var_03_vec = Vector{MyFloat}(undef, steps)
@@ -137,10 +150,18 @@ function demo_syslog(P, O=1, D=0, L=0, W=1, T=W, S=0; duration=10)
         alpha4_vec[i+1] = state.alpha4
         CL2_vec[i+1] = state.CL2
         CD2_vec[i+1] = state.CD2
-        aero_force_b_vec[i+1] = state.aero_force_b
-        aero_moment_b_vec[i+1] = state.aero_moment_b
-        tether_induced_force_vec[i+1] = state.tether_induced_force
-        tether_induced_moment_vec[i+1] = state.tether_induced_moment
+        aero_force_b_x_vec[i+1] = state.aero_force_b_x
+        aero_force_b_y_vec[i+1] = state.aero_force_b_y
+        aero_force_b_z_vec[i+1] = state.aero_force_b_z
+        aero_moment_b_x_vec[i+1] = state.aero_moment_b_x
+        aero_moment_b_y_vec[i+1] = state.aero_moment_b_y
+        aero_moment_b_z_vec[i+1] = state.aero_moment_b_z
+        tether_induced_force_b_x_vec[i+1] = state.tether_induced_force_b_x
+        tether_induced_force_b_y_vec[i+1] = state.tether_induced_force_b_y
+        tether_induced_force_b_z_vec[i+1] = state.tether_induced_force_b_z
+        tether_induced_moment_b_x_vec[i+1] = state.tether_induced_moment_b_x
+        tether_induced_moment_b_y_vec[i+1] = state.tether_induced_moment_b_y
+        tether_induced_moment_b_z_vec[i+1] = state.tether_induced_moment_b_z
         twist_angles_vec[i+1] = state.twist_angles
         vel_kite_vec[i+1] = state.vel_kite
         acc_vec[i+1] = state.acc
@@ -158,6 +179,7 @@ function demo_syslog(P, O=1, D=0, L=0, W=1, T=W, S=0; duration=10)
         drag_force_y_vec[i+1] = state.drag_force_y
         drag_force_z_vec[i+1] = state.drag_force_z
         spring_force_vec[i+1] = state.spring_force
+        gamma_distribution_vec[i+1] = state.gamma_distribution
         turn_rate_x_vec[i+1] = state.turn_rate_x
         turn_rate_y_vec[i+1] = state.turn_rate_y
         turn_rate_z_vec[i+1] = state.turn_rate_z
@@ -167,6 +189,9 @@ function demo_syslog(P, O=1, D=0, L=0, W=1, T=W, S=0; duration=10)
         set_torque_vec[i+1] = state.set_torque
         set_speed_vec[i+1] = state.set_speed
         set_force_vec[i+1] = state.set_force
+        set_ext_force_enu_x_vec[i+1] = state.set_ext_force_enu_x
+        set_ext_force_enu_y_vec[i+1] = state.set_ext_force_enu_y
+        set_ext_force_enu_z_vec[i+1] = state.set_ext_force_enu_z
         var_01_vec[i+1] = state.var_01
         var_02_vec[i+1] = state.var_02
         var_03_vec[i+1] = state.var_03
@@ -184,17 +209,19 @@ function demo_syslog(P, O=1, D=0, L=0, W=1, T=W, S=0; duration=10)
         var_15_vec[i+1] = state.var_15
         var_16_vec[i+1] = state.var_16
     end
-    StructArray{SysState{P, O, D, L, W, T, S, MyFloat}}((time_vec, t_sim_vec, sys_state_vec, cycle_vec, fig_8_vec, e_mech_vec, 
+    StructArray{SysState{P, O, D, L, W, T, S, N, MyFloat}}((time_vec, t_sim_vec, sys_state_vec, cycle_vec, fig_8_vec, e_mech_vec, 
                               Qw_vec, Qx_vec, Qy_vec, Qz_vec, turn_rates_vec, elevation_vec, 
                               azimuth_vec, azimuth_rate_vec, l_tether_vec, v_reelout_vec, winch_force_vec, depower_vec, 
                               steering_vec, kcu_steering_vec, set_steering_vec, heading_vec, heading_rate_vec, course_vec, 
                               bearing_vec, attractor_vec, v_app_vec, v_wind_gnd_vec, v_wind_200m_vec, v_wind_kite_vec, 
                               AoA_vec, side_slip_vec, alpha3_vec, alpha4_vec, CL2_vec, CD2_vec, 
-                              aero_force_b_vec, aero_moment_b_vec, tether_induced_force_vec, tether_induced_moment_vec, twist_angles_vec, vel_kite_vec, 
-                              acc_vec, X_vec, Y_vec, Z_vec, flap_angle_vec, VX_vec, 
-                              VY_vec, VZ_vec, aero_force_x_vec, aero_force_y_vec, aero_force_z_vec, drag_force_x_vec, 
-                              drag_force_y_vec, drag_force_z_vec, spring_force_vec, turn_rate_x_vec, turn_rate_y_vec, turn_rate_z_vec, 
-                              twist_vel_vec, pulley_len_vec, pulley_vel_vec, set_torque_vec, set_speed_vec, set_force_vec, 
+                              aero_force_b_x_vec, aero_force_b_y_vec, aero_force_b_z_vec, aero_moment_b_x_vec, aero_moment_b_y_vec, aero_moment_b_z_vec, 
+                              tether_induced_force_b_x_vec, tether_induced_force_b_y_vec, tether_induced_force_b_z_vec, tether_induced_moment_b_x_vec, tether_induced_moment_b_y_vec, tether_induced_moment_b_z_vec, 
+                              twist_angles_vec, vel_kite_vec, acc_vec, X_vec, Y_vec, Z_vec, 
+                              flap_angle_vec, VX_vec, VY_vec, VZ_vec, aero_force_x_vec, aero_force_y_vec, 
+                              aero_force_z_vec, drag_force_x_vec, drag_force_y_vec, drag_force_z_vec, spring_force_vec, gamma_distribution_vec, 
+                              turn_rate_x_vec, turn_rate_y_vec, turn_rate_z_vec, twist_vel_vec, pulley_len_vec, pulley_vel_vec, 
+                              set_torque_vec, set_speed_vec, set_force_vec, set_ext_force_enu_x_vec, set_ext_force_enu_y_vec, set_ext_force_enu_z_vec, 
                               var_01_vec, var_02_vec, var_03_vec, var_04_vec, var_05_vec, var_06_vec, 
                               var_07_vec, var_08_vec, var_09_vec, var_10_vec, var_11_vec, var_12_vec, 
                               var_13_vec, var_14_vec, var_15_vec, var_16_vec))